% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{barplotPetideCountPerProtein}
\alias{barplotPetideCountPerProtein}
\title{plot adj. intensity vs peptide (per protein) barplot}
\usage{
barplotPetideCountPerProtein(spectralLibrary, protLabCex = 0.9,
  acLenTrunc = 12, col = "blue", cex.axis = 1.25, cex.lab = 1.25, ...)
}
\arguments{
\item{spectralLibrary}{data.frame}

\item{protLabCex}{default 0.9}

\item{acLenTrunc}{integer default 12 "SOMEVERYLONGAC" -> "SOMEVERY.."}

\item{ptmRegExp}{default NA, higliht modified peptides in red (do not highlight labels)}
}
\description{
plot adj. intensity vs peptide (per protein) barplot
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
