% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{createComplementaryIsotopeLibrary}
\alias{createComplementaryIsotopeLibrary}
\title{Create complementary isotope (Arg, Lys H/L) spectral library}
\usage{
createComplementaryIsotopeLibrary(sl)
}
\arguments{
\item{sl}{tibble or data frame}
}
\value{
data.frame
}
\description{
Create complementary isotope (Arg, Lys H/L) spectral library
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
