% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{createLibrarySpectrum}
\alias{createLibrarySpectrum}
\title{Parse spectrum framgment match information listed in MaxQuant 'Identifications' results}
\usage{
createLibrarySpectrum(psm)
}
\arguments{
\item{psm}{a row of MaxQuant 'Identifications' data.frame}
}
\value{
list with  attributes intensity, ionType, charge, mz, isNL
}
\description{
Parse spectrum framgment match information listed in MaxQuant 'Identifications' results
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
