% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{digestProteome}
\alias{digestProteome}
\title{Get peptide candidates per protein}
\usage{
digestProteome(proteins, proteaseRegExp = "[KR](?!P)", dispProgressBar = T,
  peptideLengthRange = c(6, 21), nbMiscleavages = 0,
  exclusivePeptides = T, trimAC = T)
}
\arguments{
\item{proteins}{list}

\item{proteaseRegExp}{character '[KR](?!P)' - trypsin}

\item{dispProgressBar}{T/F}

\item{peptideLengthRange}{integer vector default c(6,21)}

\item{nbMiscleavages}{0}

\item{exclusivePeptides}{TRUE/FALSE get exclusive peptides only i.e. peptides mapping to a single protein. default T}

\item{trimAC}{TRUE/FALSE default TRUE sp|P62258|1433E_HUMAN -> P62258}
}
\value{
data.frame peptide,protein
}
\description{
Get peptide candidates per protein
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
