% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserOptions.R
\name{getCMDLineOptions}
\alias{getCMDLineOptions}
\title{Define and User Specified Command Line Options}
\usage{
getCMDLineOptions(version = version)
}
\arguments{
\item{version}{SpectroX version number}
}
\value{
list() user options list
}
\description{
Define and User Specified Command Line Options
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
