% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getEmpiricalIRT}
\alias{getEmpiricalIRT}
\title{Add iRT metric to data.frame}
\usage{
getEmpiricalIRT(tb, fit)
}
\arguments{
\item{tb}{tibble containing colum labelled "Retention time"}

\item{fit}{lm object}
}
\value{
vector of normalised rt (empirical irt)
}
\description{
Add iRT metric to data.frame
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Using iRT, a normalized retention time for more targeted measurement of peptides, Escher et al. (2012), \url{http://www.ncbi.nlm.nih.gov/pubmed/22577012}
}
