% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getFragmentSequence}
\alias{getFragmentSequence}
\title{Get Fragment Sub Sequnece}
\usage{
getFragmentSequence(peptide = peptide, ionType = ionType,
  fragmentNb = fragmentNb)
}
\arguments{
\item{peptide}{character string}

\item{ionType}{c("a","b","x","y")}

\item{fragmentNb}{integer}
}
\value{
character string
}
\description{
Get Fragment Sub Sequnece
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
