% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getIRTModel}
\alias{getIRTModel}
\title{Get linear model predicting iRT as a function retention time (column name "rt")}
\usage{
getIRTModel(tb)
}
\arguments{
\item{tb}{tibble including column "Retention time" and "Sequence"}
}
\value{
list including fit rlm object,  data.frame peptide,rt,irtRef
}
\description{
Get linear model predicting iRT as a function retention time (column name "rt")
}
\details{
Uses Robust Linear Regression. Retention times of at least 3 iRT unique peptides needs to be listed in the input tibble
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
Using iRT, a normalized retention time for more targeted measurement of peptides, Escher et al. (2012), \url{http://www.ncbi.nlm.nih.gov/pubmed/22577012}
}
