% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getLabelMzShift}
\alias{getLabelMzShift}
\title{Get Mz Shift of complementary peptide/fragment ion}
\usage{
getLabelMzShift(aaSeq, charge = 1, isHeavy = T)
}
\arguments{
\item{aaSeq}{character string}

\item{charge}{default 1}

\item{isHeavy}{c(T,F)}
}
\value{
numeric
}
\description{
Get Mz Shift of complementary peptide/fragment ion
}
\details{
Calculate mass shift of complimentary spectrum.
 \itemize{
 \item PETIDEK (light) -> 8.014199
 \item PETIDEK (heavy) -> -8.014199
 \item PETIDER (light) -> 10.008269
 \item PETIDER (heavy) -> -10.008269
 }
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
