% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getPTMColors}
\alias{getPTMColors}
\title{assign colors to ptms}
\usage{
getPTMColors(ptms)
}
\arguments{
\item{ptms}{vector}
}
\value{
data.frame
}
\description{
assign colors to ptms
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
