% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getSearchedModifications}
\alias{getSearchedModifications}
\title{Get list of variable modifiections condsidered by MaxQuant}
\usage{
getSearchedModifications(tb)
}
\arguments{
\item{df}{tible or data.frame}
}
\value{
character vector of modification names
}
\description{
Get list of variable modifiections condsidered by MaxQuant
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
