% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserOptions.R
\name{getUserOptions}
\alias{getUserOptions}
\title{check and return User Specified Command Line Options}
\usage{
getUserOptions(cmdlineOptions)
}
\arguments{
\item{list()}{command line options}
}
\value{
list() user options
}
\description{
check and return User Specified Command Line Options
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
