% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserOptions.R
\name{parseRange}
\alias{parseRange}
\title{Parse range in format 2:4 and return integer vector. Inthis case 2,3,4}
\usage{
parseRange(rangeStr)
}
\arguments{
\item{rangeStr}{e.g. '2:4'}
}
\value{
vectot of integers
}
\description{
Parse range in format 2:4 and return integer vector. Inthis case 2,3,4
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
