% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{parseTargetsFile}
\alias{parseTargetsFile}
\title{Parse list of target peptides/proteins}
\usage{
parseTargetsFile(file)
}
\arguments{
\item{file}{path}
}
\value{
list() peptides, proteins
}
\description{
Parse list of target peptides/proteins
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
