% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{plotIRTCalibration}
\alias{plotIRTCalibration}
\title{Plot IRT calibration curve}
\usage{
plotIRTCalibration(irtModel, cex = 1.5, col = "blue", pch = 19, ...)
}
\arguments{
\item{irtModel}{list}
}
\description{
Plot IRT calibration curve
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
