% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{proteotypicPeptideExport}
\alias{proteotypicPeptideExport}
\title{# Export top X peptide per protein (ordered by 'adjustedIntensitySum') and add theoretical peptides ranked by length if fewer than X peptides were identified by MaxQuant}
\usage{
proteotypicPeptideExport(spectralLibrary, theoPeptides = NA,
  targetProteins = unique(spectralLibrary$protein),
  nbPeptidesPerProtein = 5, outFile = paste0(tempdir(), "/tmp.xls"))
}
\arguments{
\item{spectralLibrary}{tibble}

\item{theoPeptides}{data.frame 'peptide' 'protein' 'length'}

\item{targetProteins}{list of protein accesion numbers e.g. P62258. default all proteins in spectralLibrary}

\item{nbPeptidesPerProtein}{number of peptides per protein to be exported}

\item{outFile}{path to output file defult tmp.xls in tmp dir}
}
\description{
# Export top X peptide per protein (ordered by 'adjustedIntensitySum') and add theoretical peptides ranked by length if fewer than X peptides were identified by MaxQuant
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
