% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{skylineExport}
\alias{skylineExport}
\title{Write skyline compatible xls file}
\usage{
skylineExport(sl, file)
}
\arguments{
\item{sl}{tibble or data frame}
}
\description{
Write skyline compatible xls file
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
