% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{spectroDiveExport}
\alias{spectroDiveExport}
\title{Write SpectroDive compatible xls file}
\usage{
spectroDiveExport(sl, file)
}
\arguments{
\item{sl}{tibble or data frame}
}
\description{
Write SpectroDive compatible xls file
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
