% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{spectronautExport}
\alias{spectronautExport}
\title{Write Spectronaut compatible xls file}
\usage{
spectronautExport(sl, file)
}
\arguments{
\item{sl}{tibble or data frame}
}
\description{
Write Spectronaut compatible xls file
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
