% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{barplotPeptidesPerProtein}
\alias{barplotPeptidesPerProtein}
\title{plot adj. intensity vs peptide (per protein) barplot}
\usage{
barplotPeptidesPerProtein(df, ptmRegExp = NA, pepLenTrunc = 12,
  pepLabCex = 0.7, ...)
}
\arguments{
\item{df}{data.frame}

\item{ptmRegExp}{default NA, higliht modified peptides in red (do not highlight labels)}

\item{pepLenTrunc}{integer AFADAMEVIPSTLAENAGLNPISTVTELR -> AFADAMEVIPSTLAE..}

\item{pepLabCex}{default 0.7}
}
\description{
plot adj. intensity vs peptide (per protein) barplot
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
