% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{barplotPeptidesPerProtein}
\alias{barplotPeptidesPerProtein}
\title{plot ranking metric vs peptide (per protein) barplot}
\usage{
barplotPeptidesPerProtein(df, pepLenTrunc = 12, pepLabCex = 0.7,
  rankingMetric = "rankingMetric", ptmCol = getPTMColors(levels(df$ptm)),
  ...)
}
\arguments{
\item{df}{data.frame}

\item{pepLenTrunc}{integer AFADAMEVIPSTLAENAGLNPISTVTELR -> AFADAMEVIPSTLAE..}

\item{pepLabCex}{default 0.7}

\item{rankingMetric}{character}

\item{ptmCol}{data.frame row.names: ptm col:col}
}
\description{
plot ranking metric vs peptide (per protein) barplot
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
