% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{barplotPetideCountPerProtein}
\alias{barplotPetideCountPerProtein}
\title{plot peptide count per protein}
\usage{
barplotPetideCountPerProtein(spectralLibrary, protLabCex = 0.9,
  acLenTrunc = 12, col = "blue", ...)
}
\arguments{
\item{spectralLibrary}{data.frame}

\item{protLabCex}{default 0.9}

\item{acLenTrunc}{integer default 12 "SOMEVERYLONGAC" -> "SOMEVERY.."}
}
\description{
plot peptide count per protein
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
