% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{getSearchedModifications}
\alias{getSearchedModifications}
\title{Get list of variable modifiections condsidered by MaxQuant}
\usage{
getSearchedModifications(tb, ignoreArgLysIsoLabel = T)
}
\arguments{
\item{ignoreArgLysIsoLabel}{default T Do not consider Arg Lys C,N heavy isotope label}

\item{df}{tible or data.frame}
}
\value{
character vector of modification names (does not return "Arg10","Lys8" labels )
}
\description{
Get list of variable modifiections condsidered by MaxQuant
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
