% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroX.R
\name{parseMaxQuantMSMS}
\alias{parseMaxQuantMSMS}
\title{Parse MaxQuant msms.txt}
\usage{
parseMaxQuantMSMS(file, pepCutoff = 0.05, targetPeptides = NA,
  targetProteins = NA, filterContaminants = T,
  contaminantRegExp = "^CON_", selectedPTMRegExp = NA,
  filterNonExclusivePeptides = T, pepLength = c(1, Inf),
  chargeState = c(1, Inf), label = NA, maxMissedCleavages = 0,
  keepBestSpectrumOnly = T, requiredPepSeqRegExp = ".", ...)
}
\arguments{
\item{file}{path}

\item{pepCutoff}{numeric default 0.05}

\item{targetPeptides}{character default NA}

\item{targetProteins}{character default NA}

\item{filterContaminants}{TRUE}

\item{contaminantRegExp}{'^CON_'}

\item{selectedPTMRegExp}{NA}

\item{filterNonExclusivePeptides}{default TRUE}

\item{pepLength}{default [1,Inf)}

\item{chargeState}{default [1,Inf)}

\item{label}{(Arg10 and Lys8 only filter) options NA (deafault,no filter),  'light','heavy'}

\item{keepBestSpectrumOnly}{keep top-scoring spectrum only, per peptidoform, default TRUE}

\item{requiredSequenceRegExp}{dafualt '.' (no filter)}
}
\value{
data.frame of maxQuant psm level results
}
\description{
Parse MaxQuant msms.txt
}
\details{
No details
}
\note{
No note
}
\examples{
print("No examples")
}
\references{
NA
}
